+++ src/main/java/com/daily/gaboja/product/dto/ProductCreateRequestDto.java
@@ -1,5 +1,6 @@
 package com.daily.gaboja.product.dto;
 
+import com.daily.gaboja.product.domain.Product;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,7 +11,7 @@ import javax.validation.constraints.PositiveOrZero;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class ProductUpdateRequest {
+public class ProductCreateRequestDto {
 
     @NotBlank(message = "상품 이름을 입력하시오.")
     private String name;
@@ -23,4 +24,13 @@ public class ProductUpdateRequest {
 
     @NotBlank(message = "상품 카테고리를 입력하시오.")
     private String category;
+
+    public Product toEntity() {
+        return Product.builder()
+                .name(name)
+                .description(description)
+                .stock(stock)
+                .category(category)
+                .build();
+    }
 }
